Move batch orientation operations
