+++ pytransform3d/trajectories.pyi
@@ -44,12 +44,6 @@ def dual_quaternions_from_transforms(A2Bs: npt.ArrayLike) -> np.ndarray: ...
 def pqs_from_dual_quaternions(dqs: npt.ArrayLike) -> np.ndarray: ...
 
 
-def norm_axis_angles(a: npt.ArrayLike) -> np.ndarray: ...
-
-
-def axis_angles_from_quaternions(qs: npt.ArrayLike) -> np.ndarray: ...
-
-
 def screw_parameters_from_dual_quaternions(
         dqs: npt.ArrayLike) -> np.ndarray: ...
